// src/app/Shared/models/life/setup/organization/appraisal/team.ts
export class Team {
  "appraisalCode": string;
  "auth": number;
  "authBy": string;
  "authOn": Date;
  "createdBy": string;
  "createdOn": Date;
  "criteria1": string;
  "criteria1Actual": number;
  "criteria1Cond": string;
  "criteria1ProdClass": string;
  "criteria1Target": number;
  "criteria1Weight": number;
  "criteria2": string;
  "criteria2Actual": number;
  "criteria2Cond": string;
  "criteria2ProdClass": string;
  "criteria2Target": number;
  "criteria2Weight": number;
  "criteria3": string;
  "criteria3Actual": number;
  "criteria3Cond": string;
  "criteria3ProdClass": string;
  "criteria3Target": number;
  "criteria3Weight": number;
  "criteria4": string;
  "criteria4Actual": number;
  "criteria4Cond": string;
  "criteria4ProdClass": string;
  "criteria4Target": number;
  "criteria4Weight": number;
  "criteria5Actual": number;
  "criteria5ProdClass": string;
  "description": string;
  "id": number;
  "indPerfWeight": number;
  "overallScore": number;
  "teamPerfWeight": number;
  "versionDate": Date;
  "versionNo": number;
}